move cross-package testing utils into base package
