+++ tools/c7n_azure/requirements.txt
@@ -1,3 +1,4 @@
 vcrpy==1.11.0
 vcrpy_unittest
 adal~=0.5.0
+backports.functools_lru_cache
